活动列表搜索;添加关注更新关注人数

+++ job-server/src/main/java/com/worldelite/job/service/ActivityService.java
@@ -4,6 +4,7 @@ import cn.hutool.core.bean.BeanUtil;
 import com.github.pagehelper.Page;
 import com.worldelite.job.constants.ActivityStatus;
 import com.worldelite.job.constants.FavoriteType;
+import com.worldelite.job.constants.UserType;
 import com.worldelite.job.entity.Activity;
 import com.worldelite.job.entity.ActivityOptions;
 import com.worldelite.job.entity.Favorite;
@@ -18,6 +19,7 @@ import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.lang.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
@@ -28,7 +30,7 @@ import java.util.List;
  * @author yeguozhong yedaxia.github.com
  */
 @Service
-public class ActivityService extends BaseService{
+public class ActivityService extends BaseService {
 
     @Autowired
     private ActivityMapper activityMapper;
@@ -44,10 +46,11 @@ public class ActivityService extends BaseService{
 
     /**
      * 获取活动列表
+     *
      * @param listForm
      * @return
      */
-    public PageResult<ActivityVo> getActivityList(ActivityListForm listForm){
+    public PageResult<ActivityVo> getActivityList(ActivityListForm listForm) {
         ActivityOptions options = new ActivityOptions();
         BeanUtil.copyProperties(listForm, options);
         options.setCityIds(StringUtils.join(listForm.getCityIds(), ","));
@@ -55,7 +58,7 @@ public class ActivityService extends BaseService{
         Page<Activity> activityPage = (Page<Activity>) activityMapper.selectAndList(options);
         PageResult<ActivityVo> pageResult = new PageResult<>(activityPage);
         List<ActivityVo> activityVoList = new ArrayList<>(activityPage.size());
-        for(Activity activity: activityPage){
+        for (Activity activity : activityPage) {
             activityVoList.add(toActivityVo(activity));
         }
         pageResult.setList(activityVoList);
@@ -68,47 +71,47 @@ public class ActivityService extends BaseService{
      * @param id
      * @return
      */
-    public ActivityVo getActivityInfo(Integer id){
+    public ActivityVo getActivityInfo(Integer id) {
         Activity activity = activityMapper.selectByPrimaryKey(id);
         return toActivityVo(activity);
     }
 
-    public ActivityVo getSimpleActivity(Integer id){
+    public ActivityVo getSimpleActivity(Integer id) {
         Activity activity = activityMapper.selectSimpleById(id);
         return toActivityVo(activity);
     }
 
-    public PageResult<ActivityVo> getSimpleActivityByStatus(Long userId,ActivityListForm pageForm){
+    public PageResult<ActivityVo> getSimpleActivityByStatus(Long userId, ActivityListForm pageForm) {
         //根据不同状态构建查询SQL
         StringBuffer where = new StringBuffer();
         Date date = new Date();
         String strDateFormat = "yyyy-MM-dd HH:mm:ss";
         SimpleDateFormat sdf = new SimpleDateFormat(strDateFormat);
         String dateStr = sdf.format(date);
-        where.append("id in (select object_id from t_favorite where type = 3 and user_id = "+userId+")");
+        where.append("id in (select object_id from t_favorite where type = 3 and user_id = " + userId + ")");
         Byte status = pageForm.getStatus();
-        if(status == ActivityStatus.WILL.value){
-            where.append(" and start_time > '"+dateStr+"'");
+        if (status == ActivityStatus.WILL.value) {
+            where.append(" and start_time > '" + dateStr + "'");
         }
-        if(status == ActivityStatus.ACTIVE.value){
-            where.append(" and start_time < '"+dateStr+"'");
-            where.append(" and finish_time > '"+dateStr+"'");
+        if (status == ActivityStatus.ACTIVE.value) {
+            where.append(" and start_time < '" + dateStr + "'");
+            where.append(" and finish_time > '" + dateStr + "'");
         }
-        if(status == ActivityStatus.END.value){
-            where.append(" and finish_time < '"+dateStr+"'");
+        if (status == ActivityStatus.END.value) {
+            where.append(" and finish_time < '" + dateStr + "'");
         }
         AppUtils.setPage(pageForm);
-        Page<Activity> page = (Page<Activity>)activityMapper.selectSimpleByIdAndStatus(where.toString());
+        Page<Activity> page = (Page<Activity>) activityMapper.selectSimpleByIdAndStatus(where.toString());
         PageResult<ActivityVo> pageResult = new PageResult<>(page);
         List<ActivityVo> activityVoList = new ArrayList<>(page.size());
-        for(Activity activity: page){
+        for (Activity activity : page) {
             ActivityVo activityVo = toActivityVo(activity);
             Favorite options = new Favorite();
             options.setType(FavoriteType.ACTIVITY.value);
             options.setUserId(userId);
             options.setObjectId(activity.getId().longValue());
             List<Favorite> favoriteList = favoriteMapper.selectAndList(options);
-            if(CollectionUtils.isNotEmpty(favoriteList)){
+            if (CollectionUtils.isNotEmpty(favoriteList)) {
                 activityVo.setJoinTime(favoriteList.get(0).getCreateTime());
             }
             activityVoList.add(activityVo);
@@ -123,21 +126,35 @@ public class ActivityService extends BaseService{
      * @param activityForm
      * @return
      */
-    public void saveActivity(ActivityForm activityForm){
+    @Transactional
+    public void saveActivity(ActivityForm activityForm) {
         Activity activity = null;
-        if(activityForm.getId() != null){
+        if (activityForm.getId() != null) {
             activity = activityMapper.selectSimpleById(activityForm.getId());
         }
 
-        if(activity == null){
-           activity = new Activity();
+        if (activity == null) {
+            activity = new Activity();
         }
         BeanUtil.copyProperties(activityForm, activity);
-        activity.setThumbnail(AppUtils.getOssKey(activityForm.getThumbnail()));
-        if(activity.getId() == null){
-            activity.setStatus(ActivityStatus.PUBLISH.value);
+
+        activity.setPoster(AppUtils.getOssKey(activityForm.getPoster()));
+
+        if(activity.getUserId() == null){
+            activity.setUserId(curUser().getId());
+            activity.setUserType(String.valueOf(curUser().getType()));
+        }
+        //TODO 需要保存组织信息
+
+        if (activity.getId() == null) {
+            //新发布的活动都是未来将举办的,状态默认即将开始
+            if (getActivityStatus(activity) == ActivityStatus.WILL)
+                activity.setStatus(ActivityStatus.WILL.value);
+            else
+                throw new RuntimeException(message("activity.status.error"));
+
             activityMapper.insertSelective(activity);
-        }else{
+        } else {
             activity.setUpdateTime(new Date());
             activityMapper.updateByPrimaryKeySelective(activity);
         }
@@ -148,11 +165,16 @@ public class ActivityService extends BaseService{
      *
      * @param id
      */
-    public void takeOffActivity(Integer id){
+    public void takeOffActivity(Integer id) {
         Activity activity = activityMapper.selectSimpleById(id);
-        if(activity != null){
-            activity.setStatus(ActivityStatus.OFFLINE.value);
-            activityMapper.updateByPrimaryKeySelective(activity);
+        if (activity != null) {
+            //自己发布的或者管理员才能下架活动
+            if(activity.getUserId().equals(curUser().getId()) || curUser().getType() == UserType.ADMIN.value) {
+                activity.setStatus(ActivityStatus.OFFLINE.value);
+                activityMapper.updateByPrimaryKeySelective(activity);
+            }else{
+                throw new RuntimeException(message("api.error.permission.denied"));
+            }
         }
     }
 
@@ -161,33 +183,62 @@ public class ActivityService extends BaseService{
      *
      * @param id
      */
-    public void deleteActivity(Integer id){
-        activityMapper.deleteByPrimaryKey(id);
+    public void deleteActivity(Integer id) {
+        Activity activity = activityMapper.selectSimpleById(id);
+        if (activity != null) {
+            //自己发布的或者管理员才能删除活动
+            if (activity.getUserId().equals(curUser().getId()) || curUser().getType() == UserType.ADMIN.value) {
+                activityMapper.deleteByPrimaryKey(id);
+            }else{
+                throw new RuntimeException(message("api.error.permission.denied"));
+            }
+        }
     }
 
-    private ActivityVo toActivityVo(Activity activity){
-        if(activity == null){
+    private ActivityVo toActivityVo(Activity activity) {
+        if (activity == null) {
             return null;
         }
         ActivityVo activityVo = new ActivityVo().asVo(activity);
-        Byte status = activityVo.getStatus();
-        Date date = new Date();
-        Date startTime = activityVo.getStartTime();
-        Date finishTime = activityVo.getFinishTime();
-        if(date.compareTo(startTime)<0){
-            activityVo.setStatus(ActivityStatus.WILL.value);
-        }
-        if(date.compareTo(startTime)>=0 && date.compareTo(startTime)<=0){
-            activityVo.setStatus(ActivityStatus.ACTIVE.value);
-        }
-        if(date.compareTo(finishTime)>0){
-            activityVo.setStatus(ActivityStatus.END.value);
-        }
+        activityVo.setStatus(getActivityStatus(activity).value);
+
         activityVo.setCurTime(new Date());
         activityVo.setCity(cityService.getCityVo(activity.getCityId()));
-        if(curUser() != null){
+        if (curUser() != null) {
             activityVo.setJoinFlag(favoriteService.checkUserFavorite(activity.getId().longValue(), FavoriteType.ACTIVITY));
         }
         return activityVo;
+    }
+
+    public ActivityStatus getActivityStatus(Activity activity) {
+        Date date = new Date();
+        ActivityStatus status;
+        if (date.compareTo(activity.getRegistrationStartTime()) < 0) {
+            status = ActivityStatus.WILL;
+        } else if (date.compareTo(activity.getRegistrationStartTime()) >= 0 && date.compareTo(activity.getRegistrationFinishTime()) <= 0) {
+            status = ActivityStatus.SIGN_UP;
+        } else if (date.compareTo(activity.getRegistrationFinishTime()) > 0 && date.compareTo(activity.getActivityStartTime()) < 0) {
+            status = ActivityStatus.WILL;
+        } else if (date.compareTo(activity.getActivityStartTime()) >= 0 && date.compareTo(activity.getActivityFinishTime()) <= 0) {
+            status = ActivityStatus.ACTIVE;
+        } else {
+            status = ActivityStatus.END;
+        }
+
+        return status;
+    }
+
+    /**
+     * 减去一个关注
+     */
+    public void minusFollower(Integer activityId) {
+        activityMapper.minusFollower(activityId);
+    }
+
+    /**
+     * 增加一个关注
+     */
+    public void increaseFollower(Integer activityId) {
+        activityMapper.increaseFollower(activityId);
     }
 }
